Corrected some bugs. Starting dokumentation.
