+++ src/enjug/erijan/games/yatzy/GameControl.java
@@ -3,29 +3,12 @@ package enjug.erijan.games.yatzy;
 import enjug.erijan.games.util.GameDie;
 
 /**
- * Created by Jan Eriksson on 16/11/15.
+ * Created by Jan Eriksson on 30/11/15.
  */
-public class GameControl {
+public interface GameControl {
+  void toggleActiveDie(GameDie die);
 
-  private Scoring scoring;
-  private RollControl rollControl;
+  void setScore(String boxId);
 
-  public GameControl(Scoring scoring, RollControl rollControl) {
-    this.scoring = scoring;
-    this.rollControl = rollControl;
-  }
-
-  public void toggleActiveDie(GameDie die) {
-    rollControl.toggleActiveDie(die);
-  }
-
-  public void setScore(String boxId) {
-    scoring.setScore(boxId);
-    rollControl.resetDice();
-  }
-
-  public void rollActiveDice() {
-    rollControl.rollActiveDice();
-    scoring.setTempScores();
-  }
+  void rollActiveDice();
 }
